split correctly all hub code inside the hub mod
