+++ src/hub/mod.rs
@@ -1,0 +1,33 @@
+//! Module for handling Hugging Face Hub interactions
+
+// Hub Struct for handling Hugging Face Hub interactions
+// Config
+mod config;
+pub use config::Config;
+// LFS info
+mod lfs_info;
+pub use lfs_info::LfsInfo;
+// Model File
+mod model_file;
+pub use model_file::ModelFile;
+// Model Info
+mod model_info;
+pub use model_info::ModelInfo;
+// Siblings
+mod siblings;
+pub use siblings::Siblings;
+
+// Hub methods for getting model info
+// Hub methods
+mod hub;
+pub use hub::{
+    list_files_info,
+    retrieve_model_info,
+};
+// Utils
+mod utils;
+pub use utils::{
+    CUSTOM_ENCODE_SET,
+    HUB_ENDPOINT,
+    build_headers,
+};
